feat(frontend): refactor
